+++ ricaun.RevitTest.Command/Process/RevitTestProcessStartUtils.cs
@@ -56,6 +56,7 @@ namespace ricaun.RevitTest.Command.Process
             string language = null,
             bool revitOpen = false,
             bool revitClose = false,
+            int timeoutMinutes = 0,
             bool debugger = false,
             Action<string> consoleAction = null,
             Action<string> debugAction = null,
@@ -69,6 +70,7 @@ namespace ricaun.RevitTest.Command.Process
                 .SetOutputConsole()
                 .SetOpen(revitOpen)
                 .SetClose(revitClose)
+                .SetTimeout(timeoutMinutes)
                 .SetLog()
                 .SetTestFilter(filter)
                 .SetDebugger(debugger)